third_party/blink/web_tests/fast/dom/HTMLImageElement/image-sizes-js-innerhtml-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS fileName(window.testdiv.children[0].currentSrc) is "image-set-4x.png"
